requirements.txt
asn1crypto==0.22.0
certifi==2017.11.5
chardet==3.0.4
cloudinary==1.8.0
confusable-homoglyphs==2.0.2
cryptography==1.9
Django==1.11.7
django-allauth==0.34.0
django-bootstrap3==9.1.0
django-cloudinary-storage==0.2.2
django-mongodb-engine==0.6.0
django-registration==2.3
django-registration-redux==1.8
djangorestframework==3.7.3
djangotoolbox==1.8.0
enum34==1.1.6
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.17
keyring==10.4.0
keyrings.alt==2.2
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.2
pycrypto==2.6.1
pygobject==3.24.1
python-apt==1.4.0b3
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
pyxdg==0.25
requests==2.18.4
requests-oauthlib==0.8.0
SecretStorage==2.3.1
six==1.11.0
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
